Move subheading_data to Dotcom.Alerts
