+++ test/dotcom/system_status/subway_test.exs
@@ -620,76 +620,6 @@ defmodule Dotcom.SystemStatus.SubwayTest do
     end
   end
 
-  describe "subheading_data" do
-    test "returns affected_stops subheading data for station closure" do
-      # Setup
-      alerts = [Alert.build(:alert_for_route)]
-      route_ids = [Faker.Util.pick(Subway.lines())]
-
-      # Exercise
-      result =
-        Subway.subheading_data(status: :station_closure, alerts: alerts, route_ids: route_ids)
-
-      # Verify
-      assert match?({:affected_stops, _}, result)
-    end
-
-    test "returns delay subheading data for delay status" do
-      # Setup
-      alerts = [Alert.build(:alert_for_route)]
-
-      # Exercise
-      result = Subway.subheading_data(status: :delay, alerts: alerts, route_ids: [])
-
-      # Verify
-      assert result == {:delay}
-    end
-
-    test "returns endpoint_stops subheading data for service_change, shuttle, single_tracking, and suspension with route_ids" do
-      # Setup
-      alerts = [Alert.build(:alert_for_route)]
-      endpoint_statuses = [:service_change, :shuttle, :single_tracking, :suspension]
-      route_ids = [Faker.Util.pick(Subway.lines())]
-      status = Faker.Util.pick(endpoint_statuses)
-
-      # Exercise
-      result = Subway.subheading_data(status: status, alerts: alerts, route_ids: route_ids)
-
-      # Verify
-      assert match?({:endpoint_stops, _}, result)
-    end
-
-    test "returns nil for statuses that don't have associated subheading text" do
-      # Setup
-      non_subheading_statuses =
-        Alerts.service_impacting_effects() --
-          [:station_closure, :delay, :service_change, :shuttle, :single_tracking, :suspension]
-
-      status = Faker.Util.pick(non_subheading_statuses)
-      alerts = [Alert.build(:alert_for_route)]
-      route_ids = [Faker.Util.pick(Subway.lines())]
-
-      # Exercise
-      result = Subway.subheading_data(status: status, alerts: alerts, route_ids: route_ids)
-
-      # Verify
-      assert result == nil
-    end
-
-    test "returns nil for non-delay statuses if route_ids is empty" do
-      # Setup
-      alerts = [Alert.build(:alert_for_route)]
-      non_delay_statuses = Alerts.service_impacting_effects() -- [:delay]
-      status = Faker.Util.pick(non_delay_statuses)
-
-      # Exercise
-      result = Subway.subheading_data(status: status, alerts: alerts, route_ids: [])
-
-      # Verify
-      assert result == nil
-    end
-  end
-
   # Returns the statuses for the given route_id and branch_id
   # collection. If no branches are specified, then returns the group
   # for the given route_id with an empty branch_ids list.
